requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
jinja2==3.1.5
sqlalchemy==2.0.38
alembic==1.14.1
psycopg[binary]>=3.1,<4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
